feature: return proper authorization responses to callers
Kessel's Relations API returns an "unauthorized" response for both when
the principal is trully unauthorized to perform an action over a
resource, and for when the resource itself does not exist in SpiceDB.

For endpoints where we want to fetch the single integration or the
integration is a path parameter which acts as a filter for any other
subcollection, we want to be able to accurately and safely respond to
our callers.

However, we cannot do it lightly: we need to make sure that only those
callers who have enough permissions are notified about a non-existing
integration.

That is the reason why, the access and existence check is performed in
various steps:

- First a regular access check is performed for the integration. In the
  case of a positive response from Kessel, we are sure that both the
  principal is authorized to perform the action and that the integration
  exists in SpiceDB.
- Should that first check fail, we need then to make sure that it is not
  related to a non-existing integration, and therefore we need to fetch
  it locally to check for its existence. If it exists, we know that it
  is a truthfully "unauthorized" response from Kessel, and we can simply
  forward that "unauthorized" response to the caller.
- If the resource is not present in our database, then we need to make
  sure that the caller has enough permissions to know that the
  integration is really not present in our database, since we do not
  want to leak that information to users who are not authorized. That is
  why we check if the caller has a workspace permission to view the
  integrations.
- When the caller does not have that workspace permission, we know they
  do not have enough permissions to know if the integration exists or
  not, so we simply return an "unauthorized" response.
- On the other hand, when the caller has that workspace permission we
  can simply return a proper "not found" response, letting the caller
  know that the specified integration simply does not exist.

RHCLOUD-35763